flights-scraper/requirements.txt
aiohttp==3.6.2
kafka-python==2.0.1